Fixed price type to double instead of int

+++ Shop.Contract/Models/Products/EditProductModel.cs
@@ -9,9 +9,9 @@ namespace Tranquiliza.Shop.Contract.Models
         public string Name { get; set; }
         public string Category { get; set; }
         public string Description { get; set; }
-        public int PurchaseCost { get; set; }
-        public int Price { get; set; }
-        public int Weight { get; set; }
+        public double PurchaseCost { get; set; }
+        public double Price { get; set; }
+        public double Weight { get; set; }
         public bool IsActive { get; set; }
     }
 }
